Add REST API gateway entry point to select which port to choose
Connections over TCP/IP are handled in a multi-threaded multi-port sever. The administrator starts the server by specifying a port number
for the server and a number of additional ports to bind on. A handler listens to the administrator-defined port number which acts as the
common interface to clients, the gateway. Once a client sends a GET request to http://localhost:<admin-def port>/api/1.0/, the gateway
returns a list of additional bounded ports and a specific port number <channel> the client should use.  Which additional port is selected
for the client (which should he use afterwards) is currently straightforward implemented, it's just round robin. However, once the client
received the port number <channel>, he can start sending requests to http://localhost:<channel>/api/1.0/.

A client calls the